requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.25.10
